Create Component for each button Relates #37
