+++ client/src/components/Buttons/Edit.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Edit = () => {
+  return (
+    <button className='edit-btn'>Edit Product</button>
+  )
+}
+export default Edit;
