+++ client/src/components/Products/Product.jsx
@@ -1,20 +1,23 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import AddToCard from '../Buttons/AddToCard';
+import Delete from '../Buttons/Delete';
+import Edit from '../Buttons/Edit';
 
-const deleteRequest = (id) => {
-  /// check it, it has error in console log
-  fetch(`/api/v1/products/product/${id}`, { method: 'DELETE' })
-    .then((res) => res.json())
-    .then((data) => console.log(data))
-    .catch((err) => console.log(err));
-};
+// const deleteRequest = (id) => {
+//   /// check it, it has error in console log
+//   fetch(`/api/v1/products/product/${id}`, { method: 'DELETE' })
+//     .then((res) => res.json())
+//     .then((data) => console.log(data))
+//     .catch((err) => console.log(err));
+// };
 
 // Add to cart local storage function
-const addToCartStorage = (props) => {
-  const products = JSON.parse(localStorage.getItem('products'));
-  products.push(props.product);
-  localStorage.setItem('products', JSON.stringify(products));
-};
+// const addToCartStorage = (props) => {
+//   const products = JSON.parse(localStorage.getItem('products'));
+//   products.push(props.product);
+//   localStorage.setItem('products', JSON.stringify(products));
+// };
 
 const Product = (props) => {
   const { id, name, price, image, description, pet_category, sub_category } =
@@ -26,14 +29,7 @@ const Product = (props) => {
 
   return (
     <div className='card'>
-      <button
-        className='delete-btn'
-        onClick={() => {
-          deleteRequest(id);
-          props.deleteHandler(id);
-        }}>
-        x
-      </button>
+      <Delete {...props}/>
       <h1 className='title'>{name}</h1>
       <p className='description'>{description}</p>
       <div className='product-image'>
@@ -44,10 +40,8 @@ const Product = (props) => {
       <p className='price'>{price}$</p>
       <p className='sub-category'>{sub_category}</p>
       <p className={`forPet ${pet_category}`}>{pet_category}</p>
-      <button className='add-to-cart' onClick={() => addToCartStorage(props)}>
-        Add to Cart
-      </button>
-      <button className='edit-btn'>Edit Product</button>
+      <AddToCard {...props}/>
+      <Edit/>
     </div>
   );
 };
